add stories
